+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 * *BREAKING* Upgrade React to v18. Refs UICIRC-966.
 * Make visible selected service points after page reloading. Refs UICIRC-972.
 * Get rid of automatic alert after request failure. UICIRC-967.
+* Change enums for notice method. UICIRC-973.
 
 ## [8.0.1](https://github.com/folio-org/ui-circulation/tree/v8.0.1) (2023-03-07)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v8.0.0...v8.0.1)